You are now able to delete posts

+++ public/js/deletePost.js
@@ -1,7 +1,10 @@
 const deletePost = async (event) => {
-    event.preventDefault();
+  event.preventDefault();
+  if (event.target.hasAttribute('data-id')) {
+    const id = event.target.getAttribute('data-id');
 
-    const response = await fetch('/api/blog/delete/:id', {
+
+    const response = await fetch(`/api/blog/delete/${id}`, {
         method: 'DELETE',
       });
 
@@ -10,7 +13,7 @@ const deletePost = async (event) => {
       } else {
         console.log('ERROR')
       }
-
+    }
 };
 
 $('#deleteButton').on('click', deletePost);
